Use job_finished_at to check if a job needs processing
This is, hopefully, slightly more robust than checking for the right job
statuses. Today, the right list is 'canceled', 'failed' and 'success'

All finished jobs need to have their retries cleared from the retry map.

Failed jobs need to be checked for spurios.

So from spuriobot's perspective, failed jobs are a subset of finished
jobs, and both need processing.

As for the changes to test data: the original came from Haskell `show`
output, which showed UTCTimes in a custom format! This broke the test. I
have repaired it by modifying the time format in the json.